In the process of adding spaceship to asteroid collision.
Changed the drawing of the spaceship from using the processing.rotation() function into 3 vertices that are then used in processing.triangle().
This means that we needed to use vectors to rotate the verices of the ship manually. So I added a vector rotation function.
This change to vertices was need because my collision funtion was circle to line. So we need the vertices to make lines.

Also added a line to circle collision function for future use.