Added a basic context manager.
This allows us to quickly switch between game 'contextes'.
For instance when the ship get's hit to display the spawn message.

Having to use the .bind method seems a bit sketchy, but works for now.
That is the question, should the view have it's own loops,
or should we have external loops that are run on a view.